# api/openapi/admin.openapi.json
{
  "components": {
    "schemas": {
...
      "DataResponsePlan": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "$ref": "#/components/schemas/Plan"
          }
        },
        "required": [
          "data"
        ],
        "type": "object"
      },
      "DataResponseProject": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "$ref": "#/components/schemas/Project"
          }
        },
        "required": [
          "data"
        ],
        "type": "object"
      },
...
      "DataResponseUser": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "$ref": "#/components/schemas/User"
          }
        },
        "required": [
          "data"
        ],
        "type": "object"
      },
      "ErrorDetail": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "type": "string"
          },
          "details": {},
          "message": {
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "type": "object"
      },
      "ErrorEnvelope": {
        "additionalProperties": false,
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "required": [
          "error"
        ],
        "type": "object"
      },
...
      "ListResponsePlan": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/Plan"
            },
            "type": "array"
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          }
        },
        "required": [
          "data",
          "meta"
        ],
        "type": "object"
      },
      "ListResponseProject": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/Project"
            },
            "type": "array"
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          }
        },
        "required": [
          "data",
          "meta"
        ],
        "type": "object"
      },
      "ListResponseUser": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/User"
            },
            "type": "array"
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          }
        },
        "required": [
          "data",
          "meta"
        ],
        "type": "object"
      },
...
      "Meta": {
        "additionalProperties": false,
        "properties": {
          "page": {
            "format": "int64",
            "type": "integer"
          },
          "per_page": {
            "format": "int64",
            "type": "integer"
          },
          "total": {
            "format": "int64",
            "type": "integer"
          },
          "total_pages": {
            "format": "int64",
            "type": "integer"
          }
        },
        "required": [
          "total",
          "page",
          "per_page",
          "total_pages"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "AdminJWT": {
        "bearerFormat": "JWT",
        "description": "ModelServer admin access token.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Typed management API used by the ModelServer dashboard. Proxy/LLM compatibility endpoints are intentionally excluded.",
    "title": "ModelServer Management API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/plans": {
      "get": {
        "operationId": "listPlans",
        "parameters": [
          {
            "description": "Page number, starting at one.",
            "explode": false,
            "in": "query",
            "name": "page",
            "schema": {
              "default": 1,
              "description": "Page number, starting at one.",
              "format": "int64",
              "minimum": 1,
              "type": "integer"
            }
          },
          {
            "description": "Number of plans returned per page.",
            "explode": false,
            "in": "query",
            "name": "per_page",
            "schema": {
              "default": 20,
              "description": "Number of plans returned per page.",
              "format": "int64",
              "maximum": 100,
              "minimum": 1,
              "type": "integer"
            }
          },
          {
            "description": "Plan field used for ordering.",
            "explode": false,
            "in": "query",
            "name": "sort",
            "schema": {
              "default": "created_at",
              "description": "Plan field used for ordering.",
              "type": "string"
            }
          },
          {
            "description": "Sort direction.",
            "explode": false,
            "in": "query",
            "name": "order",
            "schema": {
              "default": "desc",
              "description": "Sort direction.",
              "enum": [
                "asc",
                "desc"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListResponsePlan"
                }
              }
            },
            "description": "OK"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Forbidden"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "AdminJWT": []
          }
        ],
        "summary": "List plans",
        "tags": [
          "Plans"
        ],
        "x-modelserver-authz": {
          "mode": "rbac",
          "scope": "system",
          "permission": "system.plans.read",
          "superadmin": "required"
        }
      }
    },
    "/api/v1/plans/{planID}": {
      "get": {
        "operationId": "getPlan",
        "parameters": [
          {
            "description": "Plan identifier.",
            "in": "path",
            "name": "planID",
            "required": true,
            "schema": {
              "description": "Plan identifier.",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataResponsePlan"
                }
              }
            },
            "description": "OK"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "AdminJWT": []
          }
        ],
        "summary": "Get plan",
        "tags": [
          "Plans"
        ],
        "x-modelserver-authz": {
          "mode": "rbac",
          "scope": "system",
          "permission": "system.plans.read",
          "superadmin": "required"
        }
      }
    },
    "/api/v1/projects": {
      "get": {
        "description": "Lists non-archived projects visible to the current user.",
        "operationId": "listProjects",
        "parameters": [
          {
            "description": "Page number, starting at one.",
            "explode": false,
            "in": "query",
            "name": "page",
            "schema": {
              "default": 1,
              "description": "Page number, starting at one.",
              "format": "int64",
              "minimum": 1,
              "type": "integer"
            }
          },
          {
            "description": "Number of projects returned per page.",
            "explode": false,
            "in": "query",
            "name": "per_page",
            "schema": {
              "default": 20,
              "description": "Number of projects returned per page.",
              "format": "int64",
              "maximum": 100,
              "minimum": 1,
              "type": "integer"
            }
          },
          {
            "description": "Project field used for ordering.",
            "explode": false,
            "in": "query",
            "name": "sort",
            "schema": {
              "default": "created_at",
              "description": "Project field used for ordering.",
              "type": "string"
            }
          },
          {
            "description": "Sort direction.",
            "explode": false,
            "in": "query",
            "name": "order",
            "schema": {
              "default": "desc",
              "description": "Sort direction.",
              "enum": [
                "asc",
                "desc"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListResponseProject"
                }
              }
            },
            "description": "OK"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Forbidden"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "AdminJWT": []
          }
        ],
        "summary": "List projects",
        "tags": [
          "Projects"
        ],
        "x-modelserver-authz": {
          "mode": "authenticated",
          "scope": "none",
          "superadmin": "none"
        }
      }
    },
    "/api/v1/projects/{projectID}": {
      "get": {
        "operationId": "getProject",
        "parameters": [
          {
            "description": "Project identifier.",
            "in": "path",
            "name": "projectID",
            "required": true,
            "schema": {
              "description": "Project identifier.",
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataResponseProject"
                }
              }
            },
            "description": "OK"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "AdminJWT": []
          }
        ],
        "summary": "Get project",
        "tags": [
          "Projects"
        ],
        "x-modelserver-authz": {
          "mode": "rbac",
          "scope": "project",
          "permission": "project.read",
          "projectIdPathParam": "projectID",
          "superadmin": "bypass"
        }
      }
    },
...
    "/api/v1/users": {
      "get": {
        "operationId": "listUsers",
        "parameters": [
          {
            "description": "Page number, starting at one.",
            "explode": false,
            "in": "query",
            "name": "page",
            "schema": {
              "default": 1,
              "description": "Page number, starting at one.",
              "format": "int64",
              "minimum": 1,
              "type": "integer"
            }
          },
          {
            "description": "Number of users returned per page.",
            "explode": false,
            "in": "query",
            "name": "per_page",
            "schema": {
              "default": 20,
              "description": "Number of users returned per page.",
              "format": "int64",
              "maximum": 100,
              "minimum": 1,
              "type": "integer"
            }
          },
          {
            "description": "User field used for ordering.",
            "explode": false,
            "in": "query",
            "name": "sort",
            "schema": {
              "default": "created_at",
              "description": "User field used for ordering.",
              "type": "string"
            }
          },
          {
            "description": "Sort direction.",
            "explode": false,
            "in": "query",
            "name": "order",
            "schema": {
              "default": "desc",
              "description": "Sort direction.",
              "enum": [
                "asc",
                "desc"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListResponseUser"
                }
              }
            },
            "description": "OK"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Forbidden"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "AdminJWT": []
          }
        ],
        "summary": "List users",
        "tags": [
          "Users"
        ],
        "x-modelserver-authz": {
          "mode": "rbac",
          "scope": "system",
          "permission": "system.users.read",
          "superadmin": "required"
        }
      }
    },
...
    "/api/v1/users/{userID}": {
      "get": {
        "operationId": "getUser",
        "parameters": [
          {
            "description": "User identifier.",
            "in": "path",
            "name": "userID",
            "required": true,
            "schema": {
              "description": "User identifier.",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataResponseUser"
                }
              }
            },
            "description": "OK"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "AdminJWT": []
          }
        ],
        "summary": "Get user",
        "tags": [
          "Users"
        ],
        "x-modelserver-authz": {
          "mode": "rbac",
          "scope": "system",
          "permission": "system.users.read",
          "superadmin": "required"
        }
      }
    }
  },
  "servers": [
    {
      "url": "/"
    }
  ]
}